requirements.txt
django
django-filter
djangorestframework
djangorestframework-simplejwt
drf-yasg
celery
redis
gunicorn
psycopg2-binary
pillow
python-decouple